propagate TMPDIR and other secure-mode env vars
The glibc loader will automatically strip some environment variables
when executing a setuid binary. Manually propagate these variables
through to the executed command.

Also, update the man page, which was a bit out of date,  and learn the
`--help` option.

Details in the changelog.
