+++ .gitignore
@@ -1,0 +1,2 @@
+autouseradd
+autouseradd-suid
